+++ src/main/scala/org/typelevel/jawn/fs2/package.scala
@@ -1,15 +1,14 @@
+package org.typelevel.jawn
+
 import cats.ApplicativeError
 import cats.effect.Concurrent
-import fs2.{Chunk, Pipe, Pull, Stream}
-import org.typelevel.jawn.{AsyncParser, Facade, ParseException}
+import _root_.fs2.{Chunk, Pipe, Pull, Stream}
 
-/**
-  * Integrates the Jawn parser with fs2
+/** Integrates the Jawn parser with fs2
   */
-package object jawnfs2 {
+package object fs2 { self =>
 
-  /**
-    * Parses to any Jawn-supported AST using the specified Async mode.
+  /** Parses to any Jawn-supported AST using the specified Async mode.
     *
     * @param facade the Jawn facade to materialize `J`
     * @tparam J the JSON AST to return
@@ -34,8 +33,7 @@ package object jawnfs2 {
     src => go(AsyncParser[J](mode))(src).stream
   }
 
-  /**
-    * Emits individual JSON elements as they are parsed.
+  /** Emits individual JSON elements as they are parsed.
     *
     * @param facade the Jawn facade to materialize `J`
     * @tparam J the JSON AST to return
@@ -46,8 +44,7 @@ package object jawnfs2 {
       facade: Facade[J]): Pipe[F, A, J] =
     parseJson(AsyncParser.ValueStream)
 
-  /**
-    * Emits elements of an outer JSON array as they are parsed.
+  /** Emits elements of an outer JSON array as they are parsed.
     *
     * @param facade the Jawn facade to materialize `J`
     * @tparam J the JSON AST to return
@@ -58,13 +55,11 @@ package object jawnfs2 {
       facade: Facade[J]): Pipe[F, A, J] =
     parseJson(AsyncParser.UnwrapArray)
 
-  /**
-    * Suffix syntax and convenience methods for parseJson
+  /** Suffix syntax and convenience methods for parseJson
     */
   implicit class JsonStreamSyntax[F[_], O](stream: Stream[F, O]) {
 
-    /**
-      * Parses a source to any Jawn-supported AST using the specified Async mode.
+    /** Parses a source to any Jawn-supported AST using the specified Async mode.
       *
       * @param facade the Jawn facade to materialize `J`
       * @tparam J the JSON AST to return
@@ -74,20 +69,21 @@ package object jawnfs2 {
         F: ApplicativeError[F, Throwable],
         A: Absorbable[O],
         facade: Facade[J]): Stream[F, J] =
-      stream.through(jawnfs2.parseJson(mode))
+      stream.through(self.parseJson(mode))
 
-    /**
-      * Parses the source to a single JSON optional JSON value.
+    /** Parses the source to a single JSON optional JSON value.
       *
       * @param facade the Jawn facade to materialize `J`
       * @tparam J the JSON AST to return
       * @return some parsed JSON value, or None if the source is empty
       */
-    def runJsonOption[J](implicit F: Concurrent[F], A: Absorbable[O], facade: Facade[J]): F[Option[J]] =
+    def runJsonOption[J](implicit
+        F: Concurrent[F],
+        A: Absorbable[O],
+        facade: Facade[J]): F[Option[J]] =
       stream.parseJson(AsyncParser.SingleValue).compile.last
 
-    /**
-      * Emits individual JSON elements as they are parsed.
+    /** Emits individual JSON elements as they are parsed.
       *
       * @param facade the Jawn facade to materialize `J`
       * @tparam J the JSON AST to return
@@ -96,10 +92,9 @@ package object jawnfs2 {
         F: ApplicativeError[F, Throwable],
         A: Absorbable[O],
         facade: Facade[J]): Stream[F, J] =
-      stream.through(jawnfs2.parseJsonStream)
+      stream.through(self.parseJsonStream)
 
-    /**
-      * Emits elements of an outer JSON array as they are parsed.
+    /** Emits elements of an outer JSON array as they are parsed.
       *
       * @param facade the Jawn facade to materialize `J`
       * @tparam J the JSON AST to return
@@ -108,6 +103,6 @@ package object jawnfs2 {
         F: ApplicativeError[F, Throwable],
         A: Absorbable[O],
         facade: Facade[J]): Stream[F, J] =
-      stream.through(jawnfs2.unwrapJsonArray)
+      stream.through(self.unwrapJsonArray)
   }
 }
